hw6-OAuth/requirements.txt
Flask==2.1.0
google-cloud-datastore==2.7.1
json2html==1.3.0